+++ pyntel4004/src/hardware/suboperations/__init__.py
@@ -1,1 +1,1 @@
-"""i4004 Processor infrastructure."""
+"""i4004 Suboperations"""