+++ pyntel4004/test/03_idx/test_0200_048_instruction_fin_x.py
@@ -11,7 +11,7 @@ import pytest  # noqa
 
 
 from hardware.processor import Processor  # noqa
-from hardware.suboperation import convert_decimal_to_n_bit_slices, \
+from hardware.suboperations.utility import convert_decimal_to_n_bit_slices, \
      decimal_to_binary  # noqa
 from hardware.suboperations.registers import insert_register   # noqa