+++ pyntel4004/test/06_io_ram/test_0200_224_instruction_wrm.py
@@ -12,7 +12,7 @@ import pickle  # noqa
 import pytest  # noqa
 
 from hardware.processor import Processor  # noqa
-from hardware.suboperation import convert_to_absolute_address  # noqa
+from hardware.suboperations.utility import convert_to_absolute_address  # noqa
 from utils import encode_command_register  # noqa