+++ pyntel4004/test/06_io_ram/test_0200_232_instruction_sbm.py
@@ -11,7 +11,7 @@ import pickle  # noqa
 import pytest  # noqa
 
 from hardware.processor import Processor  # noqa
-from hardware.suboperation import convert_to_absolute_address, \
+from hardware.suboperations.utility import convert_to_absolute_address, \
     ones_complement  # noqa
 from utils import encode_command_register  # noqa