+++ pyntel4004/test/07_subroutine/test_0200_192_instruction_bbl_x.py
@@ -11,7 +11,7 @@ import pytest  # noqa
 
 
 from hardware.processor import Processor  # noqa
-from hardware.suboperation import decimal_to_binary  # noqa
+from hardware.suboperations.utility import decimal_to_binary  # noqa
 from hardware.suboperations.stack import write_to_stack  # noqa
 from hardware.exceptions import ValueOutOfRangeForStack  # noqa