+++ pyntel4004/test/08_transfer_control/test_0200_049_instruction_jin_x.py
@@ -13,7 +13,7 @@ import pytest  # noqa
 from hardware.exceptions import ProgramCounterOutOfBounds  # noqa
 from hardware.processor import Processor  # noqa
 
-from hardware.suboperation import decimal_to_binary  # noqa
+from hardware.suboperations.utility import decimal_to_binary  # noqa
 
 
 @pytest.mark.parametrize("value", [0, 1, 2, 3, 4, 5, 6, 7])