+++ pyntel4004/test/09_memory_select/test_0200_033_instruction_src_x.py
@@ -11,7 +11,7 @@ import pytest  # noqa
 
 
 from hardware.processor import Processor  # noqa
-from hardware.suboperation import decimal_to_binary  # noqa
+from hardware.suboperations.utility import decimal_to_binary  # noqa
 from hardware.suboperations.registers import insert_register, insert_registerpair  # noqa
 from hardware.exceptions import InvalidRegisterPair  # noqa